pyproject.toml
[tool.poetry]
name = "my-github"
version = "0.1.0"
description = ""
authors = ["Patrick Rachford <prachford@icloud.com>"]
readme = "README.md"
packages = [{include = "my_github"}]

[tool.poetry.dependencies]
python = "^3.11"
PyGithub = "^1.58.0"
temporalio = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"


[tool.poe.tasks]
format = [{cmd = "black ."}, {cmd = "isort ."}]
lint = [{cmd = "black --check ."}, {cmd = "isort --check-only ."}]
